move HomeController, NewProductsController, TrendingProductsController into home folder inside controller package
