+++ src/main/webapp/assets/css/home.css
@@ -284,8 +284,10 @@
 }
 
 .image--tag:hover .action__bar{
-    bottom: 43px;
-    opacity: 70%;
+    /*bottom: 50%;*/
+    opacity: 80%;
+    height: 50px;
+    /*box-sizing: border-box;*/
 }
 
 .product__tag{
@@ -306,18 +308,20 @@
     width: 100%;
     position: absolute;
     bottom: 0;
-    left: 50%;
+    left: 0;
     display: flex;
     justify-content: center;
     border-radius: 5px;
     box-shadow: 0 5px 10px rgba(0, 0, 0, 0.05);
+    height: 0;
     opacity: 0;
-    transform: translate3d(-50%, 100%, 0);
+    /*transform: translate3d(-50%, 100%, 0);*/
     transition: all 0.3s ease;
 }
 
 .action__bar button, .action__bar .see__detail{
-    padding: 14px 0px;
+    /*padding: 14px 0px;*/
+    line-height: 16px;
     flex: 1;
     font-size: 12px;
     font-weight: 500;
@@ -325,10 +329,14 @@
     background-color: var(--blue-color-medium);
     text-align: center;
     cursor: pointer;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 4px;
 }
 
 .action__bar button:hover, .action__bar .see__detail:hover{
-    opacity: 1;
+    /*opacity: 1;*/
     background-color: var(--blue-color-high);
 }
